requirements.txt
Flask==2.1.1
gunicorn==20.1.0
google-cloud-storage==1.44.0
Pillow==9.0.1
numpy==1.22.3
tensorflow==2.7.0